refactor: move client functions from main to client
